+++ tests/tests.py
@@ -1,13 +1,32 @@
-from typing import List, Type
+import os
+import sys
+import shutil
+import tempfile
+
+from typing import List, Type, Optional
 from io import StringIO
 
 from django.test import SimpleTestCase, override_settings  # type: ignore
 
 import sms
 from sms import send_sms
-from sms.backends import dummy, locmem
+from sms.backends import dummy, locmem, filebased
 from sms.backends.base import BaseSmsBackend
 from sms.message import Message
+from sms.utils import message_from_bytes, message_from_binary_file
+
+
+class BaseSmsBackendTests:
+    sms_backend: Optional[str] = None
+
+    def setUp(self) -> None:
+        self.settings_override = override_settings(
+            SMS_BACKEND=self.sms_backend
+        )
+        self.settings_override.enable()
+
+    def tearDown(self) -> None:
+        self.settings_override.disable()
 
 
 class SmsTests(SimpleTestCase):
@@ -25,11 +44,30 @@ class SmsTests(SimpleTestCase):
         )
 
     def test_backend_arg(self) -> None:
-        """Test backend argument of mail.get_connection()."""
+        """Test backend argument of sms.get_connection()."""
         self.assertIsInstance(
             sms.get_connection('sms.backends.dummy.SmsBackend'),
             dummy.SmsBackend
         )
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            self.assertIsInstance(
+                sms.get_connection(
+                    'sms.backends.filebased.SmsBackend',
+                    file_path=tmp_dir
+                ),
+                filebased.SmsBackend
+            )
+        if sys.platform == 'win32':
+            msg = ('_getfullpathname: path should be string, bytes or '
+                   'os.PathLike, not object')
+        else:
+            msg = 'expected str, bytes or os.PathLike object, not object'
+        with self.assertRaisesMessage(TypeError, msg):
+            sms.get_connection(
+                'sms.backends.filebased.SmsBackend',
+                file_path=object()
+            )
+        self.assertIsInstance(sms.get_connection(), locmem.SmsBackend)
 
     def test_custom_backend(self) -> None:
         """Test cutoms backend defined in this suite."""
@@ -50,7 +88,7 @@ class SmsTests(SimpleTestCase):
         self.assertIsInstance(sms.outbox[0].to, list)  # type: ignore
 
 
-class LocmemBackendTests(SimpleTestCase):
+class LocmemBackendTests(BaseSmsBackendTests, SimpleTestCase):
     sms_backend: str = 'sms.backends.locmem.SmsBackend'
 
     def flush_mailbox(self) -> None:
@@ -74,7 +112,7 @@ class LocmemBackendTests(SimpleTestCase):
         self.assertEqual(len(sms.outbox), 2)  # type: ignore
 
 
-class ConsoleBackendTests(SimpleTestCase):
+class ConsoleBackendTests(BaseSmsBackendTests, SimpleTestCase):
     sms_backend: str = 'sms.backends.console.SmsBackend'
 
     def test_console_stream_kwarg(self) -> None:
@@ -90,3 +128,50 @@ class ConsoleBackendTests(SimpleTestCase):
         connection.send_messages([message])  # type: ignore
         messages = stream.getvalue().split('\n' + ('-' * 79) + '\n')
         self.assertIn('from: ', messages[0])
+
+
+class FileBasedBackendTests(BaseSmsBackendTests, SimpleTestCase):
+    sms_backend = 'sms.backends.filebased.SmsBackend'
+
+    def setUp(self) -> None:
+        super().setUp()
+        self.tmp_dir = self.mkdtemp()
+        self.addCleanup(shutil.rmtree, self.tmp_dir)
+        self._settings_override = override_settings(SMS_FILE_PATH=self.tmp_dir)
+        self._settings_override.enable()
+
+    def tearDown(self) -> None:
+        self._settings_override.disable()
+        super().tearDown()
+
+    def mkdtemp(self) -> str:
+        return tempfile.mkdtemp()
+
+    def flush_mailbox(self) -> None:
+        for filename in os.listdir(self.tmp_dir):
+            os.unlink(os.path.join(self.tmp_dir, filename))
+
+    def get_mailbox_content(self) -> List[Message]:
+        messages: List[Message] = []
+        for filename in os.listdir(self.tmp_dir):
+            with open(os.path.join(self.tmp_dir, filename), 'rb') as fp:
+                session = fp.read().split(b'\n' + (b'-' * 79) + b'\n')
+            messages.extend(message_from_bytes(m) for m in session if m)
+        return messages
+
+    def test_file_sessions(self) -> None:
+        """Make sure opening a connection creates a new file"""
+        message = Message(
+            'Here is the message',
+            '+12065550100',
+            ['+441134960000']
+        )
+        connection = sms.get_connection()
+        connection.send_messages([message])  # type: ignore
+
+        self.assertEqual(len(os.listdir(self.tmp_dir)), 1)
+        tmp_file = os.path.join(self.tmp_dir, os.listdir(self.tmp_dir)[0])
+        with open(tmp_file, 'rb') as fp:
+            message = message_from_binary_file(fp)
+        self.assertEqual(message.from_phone, '+12065550100')
+        self.assertEqual(message.to, ['+441134960000'])